Various fixes
- Properly exports VERSION variable

The export line was not migrated to the main makefile as it should.

- Fixes STARK_BUILD_DIR variable

It's value would change depending on the fist call to the variable. The
correct way to define this variable is by using the `:=` operator.

- Fixes STARK_CACHE_DIR

It now contains the full path to the cache directory.

- Sets default shell to bash

This solves an issue in which that different distributions uses
different aliases to sh.

- Fixes utilization of golangci-lint
- Fixes installation of golang tools

Tools that could be installed using `go install` now are installed using
an external script.

- Added all tools by default.
- Fixes slack module

Target was renamed to be more semantic and extra variables not used
directly by the target were droped.