Fixes auto-update not working
After a run of `stark-build-sync` it would looses the remote tracking
branch and the command `stark-build-update` would stop working because it
uses a simple `git pull`.

Now `stark-build-update` uses `git pull origin main` for updating.

After the integration tests it also became clear that the sync command
should be part of project's Makefile, because on the first run the
stark-build directory is empty.

The `stark-build-sync` was moved to the README and it's now part of the
installation instructions. Also the main `Makefile` is now included
using `-include`. This will avoid include falling when run for the first
time.