Fixes strange behavior with docker base image name (#9)
The default value for DOCKER_BASE_IMAGE was the current directory of the
project. But because the project can be cloned in different directories
this value was not consistent.

The default value for DOCKER_BASE_IMAGE is now GO_PROJECT. This is a
safe assumption because we all our projects a go projects ;). Also the
docker module uses require-DOCKER_BASE_IMAGE to ensure the value is
properly set before using it. This way the value will be consistent or
it will fail.