Fix cloud function publishing
The cloud function path in GCS was broken. It was using PROJECT instead
of GO_PROJECT to compose the path.

The value of CLOUDFUNCTIONS_DEPLOY_DIR was not making sense either
because it was using a path inside stark-build instead of the project.
It was changed $(PROJECT_DIR)/tooling/terraform/ but we should revisit
this in the future.