Fix updating go tools not working (#14)
In some rare cases, the updating was failing to check that the current
version is not the desired one. This was caused by the grep command and
was processing the query string as a regular expression. So the dots in
the version string were matching any character.

The particular case we found was on golangci-lint. We were trying to
update to version `1.48.0` but is was matching the timestamp `11:48:06`

Example of the golangci-lint version that matches `1.48.0`:

```
golangci-lint has version 1.37.1 built from b39dbcd6 on 2021-02-20T11:48:06Z
```