Improves escaping on 'slack-send-message' need ENVs
Problem:

The 'GIT_LOG' variable being created was vulnerable to pieces of commit
messages like this:

 - Let's break the commit (message). "Oh great..." you think.

The problem is that the use of the message within 'slack-send-message'
is wrapping the results with single quotes, which we cannot escape in
command line. When finding the first single quote, the command would
break into default behavior and then emit a syntax error upon finding
the '(' character.

 - syntax error near unexpected token `('

Solution:

Change escaping to use double quotes instead of single quotes. This
change resulted in invalid json to be sent to Slack, which was fixed by
adding a higher level of double quotes escaping when parsing the commit
into 'GIT_LOG'