Fix slack-send-message
With the change in the previous commit of the single quote to double
quote, the shell started processing the GIT_LOG string. This means it
started also to process back quotes (`), which were not being escaped.
This would cause the shell to process the content of back quotes as
shell commands.

The back quotes were properly escaped and now the GIT_LOG variable is
processed as plain string. Also the curl was changed from --data to
--data-binary to avoid it processing the body.