Improves escaping on 'slack-send-message' need ENVs (#15)
Improves escaping on 'slack-send-message' need ENVs

Problem:

The 'GIT_LOG' variable being created was vulnerable to pieces of commit
messages like this:

 - Let's break the commit (message). "Oh great..." you think.

The problem is that the use of the message within 'slack-send-message'
is wrapping the results with single quotes, which we cannot escape in
command line. When finding the first single quote, the command would
break into default behavior and then emit a syntax error upon finding
the '(' character.

 - syntax error near unexpected token `('

Solution:

Change escaping to use double quotes instead of single quotes. This
change resulted in invalid json to be sent to Slack, which was fixed by
adding a higher level of double quotes escaping when parsing the commit
into 'GIT_LOG'

With the change explaioned above, of the single quote to double
quote, the shell started processing the GIT_LOG string. This means it
started also to process back quotes (`), which were not being escaped.
This would cause the shell to process the content of back quotes as
shell commands.

The back quotes were properly escaped and now the GIT_LOG variable is
processed as plain string. Also the curl was changed from --data to
--data-binary to avoid it processing the body.

Co-authored-by: Rodrigo Fonseca <rodrigo.fonseca@arquivei.com.br>