Adds post-merge git hook
It often occurs that after pulling new code, the go.sum is changed and
everything breaks because vendor directory is inconsistent.

Adds a new post merge hook that will call  `make vendor` after a
successful merge/pull. This was done following the same architecture as
the previous hook, so if can be overwritten by the new variable
GIT_HOOKS_POSTMERGE.