Adds a more generic docker support
The previous docker module depends heavily on googles container
registry. This commit introduces the DOCKER_IMAGE_PREFIX variable that is
supposed to replace the previous behavior. Now images are named just as
PREFIX/CMD.

Previous behavior can still be achieved by setting the GCP_PROJECT and
DOCKER_BASE_IMAGE variables. This is expected to be removed in the
future and whoever is calling the make target should set the expected
prefix.

This commit also forces the images to always be created with the full
name. So new the publish targets are not compatible with old images.
This particular case is not supposed to happen and should be considered
an error.

Fixed an error in help module that prevented targets with the slash
character to be shown.