Fix use of grep on alpine
When running the stark-build inside an alpine container the error "grep:
unrecognized option: quiet" will pop when the go module is checking the
required tools.

This happens because the script uses grep with the "--quiet" flag but on
alpine the grep don't allow for long flags and the short "-q" should be
used.

This commit changes all long flags used in grep commands for it's short
counterpart so that it is compatible with typical grep and the BusyBox's
grep that is used inside alpine.