Fixes swagger tool
The tool was broken in two different ways:

1. A variable was named wrongly causing the update command to never
update an existing tool.
2. The swagger binary is not working correctly with new versions of go,
generating an incomplete spec.

The variable was fixes and using go-get to install swagger seems to fix
the problem with the spec.

The install script had to be updated to support the swagger binary. It
differs from previous binaries because it uses 'version' instead of
'-version' for retrieving the current binary version. Now this flag if
configurable and the Makefile passes the correct form for each tool.

There is already an open issue for this problem:

https://github.com/go-swagger/go-swagger/issues/2897